Moved symbols functions into seperate file.
